+++ src/number/minNumber.ts
@@ -1,16 +1,17 @@
 import type { YobtaSyncRule } from '../ruleYobta/index.js'
 import { ruleYobta } from '../ruleYobta/index.js'
 
-export const maxMessage = (limit: number): string =>
-  `It should be within ${limit}`
+export const minNumberMessage = (limit: number): string =>
+  `It should be at least ${limit}`
 
-export const max = (
+export const minNumber = (
   limit: () => number,
-  message = maxMessage,
+  message = minNumberMessage,
 ): YobtaSyncRule<number, number> =>
   ruleYobta<number, number>(input => {
-    if (input > limit()) {
+    if (input < limit()) {
       throw new Error(message(limit()))
     }
+
     return input
   })